+++ src/LaravelModulesServiceProvider.php
@@ -2,37 +2,17 @@
 
 namespace Nwidart\Modules;
 
-use Illuminate\Support\ServiceProvider;
-use Nwidart\Modules\Providers\BootstrapServiceProvider;
-use Nwidart\Modules\Providers\ConsoleServiceProvider;
-use Nwidart\Modules\Providers\ContractsServiceProvider;
 use Nwidart\Modules\Support\Stub;
 
-class LaravelModulesServiceProvider extends ServiceProvider
+class LaravelModulesServiceProvider extends ModulesServiceProvider
 {
-    /**
-     * Indicates if loading of the provider is deferred.
-     *
-     * @var bool
-     */
-    protected $defer = false;
-
     /**
      * Booting the package.
      */
     public function boot()
     {
         $this->registerNamespaces();
-
         $this->registerModules();
-    }
-
-    /**
-     * Register all modules.
-     */
-    protected function registerModules()
-    {
-        $this->app->register(BootstrapServiceProvider::class);
     }
 
     /**
@@ -50,55 +30,12 @@ class LaravelModulesServiceProvider extends ServiceProvider
      */
     public function setupStubPath()
     {
-        $this->app->booted(function ($app) {
-            Stub::setBasePath(__DIR__ . '/Commands/stubs');
+        Stub::setBasePath(__DIR__ . '/Commands/stubs');
 
+        $this->app->booted(function ($app) {
             if ($app['modules']->config('stubs.enabled') === true) {
                 Stub::setBasePath($app['modules']->config('stubs.path'));
             }
         });
-    }
-
-    /**
-     * Register package's namespaces.
-     */
-    protected function registerNamespaces()
-    {
-        $configPath = __DIR__ . '/../config/config.php';
-        $this->mergeConfigFrom($configPath, 'modules');
-        $this->publishes([
-            $configPath => config_path('modules.php'),
-        ], 'config');
-    }
-
-    /**
-     * Register the service provider.
-     */
-    protected function registerServices()
-    {
-        $this->app->singleton('modules', function ($app) {
-            $path = $app['config']->get('modules.paths.modules');
-
-            return new Repository($app, $path);
-        });
-    }
-
-    /**
-     * Get the services provided by the provider.
-     *
-     * @return array
-     */
-    public function provides()
-    {
-        return ['modules'];
-    }
-
-    /**
-     * Register providers.
-     */
-    protected function registerProviders()
-    {
-        $this->app->register(ConsoleServiceProvider::class);
-        $this->app->register(ContractsServiceProvider::class);
     }
 }
