+++ src/LumenModulesServiceProvider.php
@@ -1,0 +1,39 @@
+<?php
+
+namespace Nwidart\Modules;
+
+use Nwidart\Modules\Support\Stub;
+
+class LumenModulesServiceProvider extends ModulesServiceProvider
+{
+    /**
+     * Booting the package.
+     */
+    public function boot()
+    {
+        $this->setupStubPath();
+    }
+
+    /**
+     * Register all modules.
+     */
+    public function register()
+    {
+        $this->registerNamespaces();
+        $this->registerServices();
+        $this->registerModules();
+        $this->registerProviders();
+    }
+
+    /**
+     * Setup stub path.
+     */
+    public function setupStubPath()
+    {
+        Stub::setBasePath(__DIR__ . '/Commands/stubs');
+
+        if (app('modules')->config('stubs.enabled') === true) {
+            Stub::setBasePath(app('modules')->config('stubs.path'));
+        }
+    }
+}
